+++ jwst/tests_nightly/general/miri/test_miri_steps_single.py
@@ -171,7 +171,7 @@ class TestMIRIWCSIFU(BaseJWSTTest):
         # Get the region file
         region = RegionsModel(crds_client.get_reference_file(result, 'regions'))
         # Choose the same plane as in the miri.py file (hardcoded for now).
-        regions = region.regions[4, :, :]
+        regions = region.regions[7, :, :]
 
         # inputs
         x, y = grid_from_bounding_box(result.meta.wcs.bounding_box)